INTEGRATE_TAILOR_APP: Fix bugs, rename, move action to user action only

+++ app/src/main/java/com/future/tailormade/core/model/request/checkout/CheckoutMeasurementRequest.kt
@@ -2,13 +2,13 @@ package com.future.tailormade.core.model.request.checkout
 
 data class CheckoutMeasurementRequest(
 		
-    val chest: Double,
+    val chest: Float,
 
-    val hips: Double,
+    val hips: Float,
 
-    val inseam: Double,
+    val inseam: Float,
 
-    val neckToWaist: Double,
+    val neckToWaist: Float,
 
-    val waist: Double
+    val waist: Float
 )
